+++ includes/Plugin.php
@@ -44,6 +44,9 @@ final class Plugin
         $this->backend_registry = new Backend_Registry();
         $peertube_upload_policy = new PeerTube_Upload_Policy_Store($this->backend_registry);
         $video_publishing_defaults = new Video_Publishing_Defaults_Store($this->backend_registry);
+        $video_block_editor_service = new Video_Block_Editor_Service($this->backend_registry, $video_publishing_defaults);
+        $video_block_editor_rest = new Video_Block_Editor_Rest($video_block_editor_service);
+        $video_block = new Video_Block();
         $peertube_publication_catalogs = new PeerTube_Publication_Catalog_Store();
         $peertube_api_factory = static fn (string $origin): PeerTube_Api_Client =>
             new PeerTube_Api_Client(new PeerTube_Http_Client($origin));
@@ -58,6 +61,8 @@ final class Plugin
         add_action('plugins_loaded', array(Model_Activator::class, 'maybe_upgrade'));
         add_action('init', array(Video_Post_Type::class, 'register'), 5);
         add_action('init', array(Video_Meta::class, 'register'), 6);
+        add_action('init', array($video_block, 'register'), 7);
+        add_action('rest_api_init', array($video_block_editor_rest, 'register'));
         add_action('init', array(Activator::class, 'schedule_dispatch'));
         add_action('add_attachment', array($queue, 'maybe_enqueue_attachment'));
         add_action('delete_attachment', array($queue, 'delete_attachment'));